Separated Types utilities into SymPath.hs and TypesToC.hs
