+++ CarpHask.cabal
@@ -28,7 +28,9 @@ library
                        Commands,
                        Template,
                        ToTemplate,
+                       SymPath,
                        Types,
+                       TypesToC,
                        Util,
                        Eval,
                        TypeError,
