+++ src/ArrayTemplates.hs
@@ -6,6 +6,7 @@ import Debug.Trace
 
 import Util
 import Types
+import TypesToC
 import Obj
 import Parsing
 import Template
