+++ src/Deftype.hs
@@ -8,6 +8,7 @@ import Debug.Trace
 
 import Obj
 import Types
+import TypesToC
 import Util
 import Template
 import ToTemplate
