+++ src/Sumtypes.hs
@@ -6,6 +6,7 @@ import Debug.Trace
 
 import Obj
 import Types
+import TypesToC
 import Util
 import Concretize
 import Polymorphism
