+++ src/Types.hs
@@ -1,29 +1,28 @@
 module Types ( TypeMappings
              , Ty(..)
              , showMaybeTy
-             , tyToC
-             , tyToCLambdaFix
-             , tyToCRawFunctionPtrFix
              , isTypeGeneric
-             , SymPath(..)
              , unifySignatures
              , replaceTyVars
-             , mangle
-             , pathToC
              , areUnifiable
              , typesDeleterFunctionType
              , typesCopyFunctionType
              , isFullyGenericType
-             , consPath
              , doesTypeContainTyVarWithName
              , lambdaEnvTy
              , typeEqIgnoreLifetimes
              , checkKinds
+             -- SymPath imports
+             , SymPath (..)
+             , mangle
+             , pathToC
+             , consPath
              ) where
 
 import qualified Data.Map as Map
 import Data.Maybe (fromMaybe)
 import Util
+import SymPath
 --import Debug.Trace
 
 -- | Carp types.
@@ -123,43 +122,6 @@ showMaybeTy :: Maybe Ty -> String
 showMaybeTy (Just t) = show t
 showMaybeTy Nothing  = "(missing-type)"
 
-tyToC :: Ty -> String
-tyToC = tyToCManglePtr False
-
-tyToCLambdaFix :: Ty -> String
-tyToCLambdaFix t@FuncTy{} = "Lambda"
-tyToCLambdaFix (RefTy FuncTy{} _) = "Lambda*"
-tyToCLambdaFix (RefTy (RefTy FuncTy{} _) _) = "Lambda**"
-tyToCLambdaFix (RefTy (RefTy (RefTy FuncTy{} _) _) _) = "Lambda***" -- | TODO: More cases needed?! What's a better way to do it..?
-tyToCLambdaFix t = tyToCManglePtr False t
-
-tyToCRawFunctionPtrFix :: Ty -> String
-tyToCRawFunctionPtrFix t@FuncTy{} = "void*"
-tyToCRawFunctionPtrFix t = tyToCManglePtr False t
-
-tyToCManglePtr :: Bool -> Ty -> String
-tyToCManglePtr _ IntTy                   = "int"
-tyToCManglePtr _ BoolTy                  = "bool"
-tyToCManglePtr _ FloatTy                 = "float"
-tyToCManglePtr _ DoubleTy                = "double"
-tyToCManglePtr _ LongTy                  = "Long"
-tyToCManglePtr _ ByteTy                  = "uint8_t"
-tyToCManglePtr _ StringTy                = "String"
-tyToCManglePtr _ PatternTy               = "Pattern"
-tyToCManglePtr _ CharTy                  = "Char"
-tyToCManglePtr _ UnitTy                  = "void"
-tyToCManglePtr _ (VarTy x)               = x
-tyToCManglePtr _ (FuncTy argTys retTy _) = "Fn__" ++ joinWithUnderscore (map (tyToCManglePtr True) argTys) ++ "_" ++ tyToCManglePtr True retTy
-tyToCManglePtr _ ModuleTy                = error "Can't emit module type."
-tyToCManglePtr b (PointerTy p)           = tyToCManglePtr b p ++ (if b then mangle "*" else "*")
-tyToCManglePtr b (RefTy r _)             = tyToCManglePtr b r ++ (if b then mangle "*" else "*")
-tyToCManglePtr _ (StructTy s [])         = tyToCManglePtr False s
-tyToCManglePtr _ (StructTy s typeArgs)   = tyToCManglePtr False s ++ "__" ++ joinWithUnderscore (map (tyToCManglePtr True) typeArgs)
-tyToCManglePtr _ (ConcreteNameTy name)   = mangle name
-tyToCManglePtr _ TypeTy                  = error "Can't emit the type of types."
-tyToCManglePtr _ MacroTy                 = error "Can't emit the type of macros."
-tyToCManglePtr _ DynamicTy               = error "Can't emit the type of dynamic functions."
-
 isTypeGeneric :: Ty -> Bool
 isTypeGeneric (VarTy _) = True
 isTypeGeneric (FuncTy argTys retTy _) = any isTypeGeneric argTys || isTypeGeneric retTy
@@ -182,69 +144,6 @@ doesTypeContainTyVarWithName _ _ = False
 
 -- | Map type variable names to actual types, eg. t0 => Int, t1 => Float
 type TypeMappings = Map.Map String Ty
-
--- | The path to a binding
-data SymPath = SymPath [String] String deriving (Ord, Eq)
-
-instance Show SymPath where
-  show (SymPath modulePath symName) =
-    if null modulePath
-    then symName
-    else joinWithPeriod modulePath ++ "." ++ symName
-
-pathToC :: SymPath -> String
-pathToC (SymPath modulePath name) =
-  concatMap ((++ "_") . mangle) modulePath ++ mangle name
-
--- | Add qualifying strings to beginning of a path.
-consPath :: [String] -> SymPath -> SymPath
-consPath qualifyers (SymPath stringPaths name) =
-  SymPath (qualifyers ++ stringPaths) name
-
--- | Replaces symbols not allowed in C-identifiers.
-mangle :: String -> String
-mangle = sreplace . creplace
-  where creplace = replaceChars (Map.fromList [('+', "_PLUS_")
-                                               ,('-', "_MINUS_")
-                                               ,('*', "_MUL_")
-                                               ,('/', "_DIV_")
-                                               ,('<', "_LT_")
-                                               ,('>', "_GT_")
-                                               ,('?', "_QMARK_")
-                                               ,('!', "_BANG_")
-                                               ,('=', "_EQ_")])
-        sreplace = replaceStrings (Map.fromList [("auto", "_AUTO_")
-                                                 ,("break", "_BREAK_")
-                                                 ,("case", "_CASE_")
-                                                 ,("const", "_CONST_")
-                                                 ,("char", "_CHAR_")
-                                                 ,("continue", "_CONTINUE_")
-                                                 ,("default", "_DEFAULT_")
-                                                 ,("do", "_DO_")
-                                                 ,("double", "_DOUBLE_")
-                                                 ,("else", "_ELSE_")
-                                                 ,("enum", "_ENUM_")
-                                                 ,("extern", "_EXTERN")
-                                                 ,("float", "_FLOAT_")
-                                                 ,("for", "_FOR")
-                                                 ,("goto", "_GOTO_")
-                                                 ,("if", "_IF_")
-                                                 ,("int", "_INT_")
-                                                 ,("long", "_LONG_")
-                                                 ,("register", "_REGISTER_")
-                                                 ,("return", "_RETURN_")
-                                                 ,("short", "_SHORT_")
-                                                 ,("signed", "_SIGNED_")
-                                                 ,("sizeof", "_SIZEOF_")
-                                                 ,("static", "_STATIC_")
-                                                 ,("struct", "_STRUCT_")
-                                                 ,("switch", "_SWITCH_")
-                                                 ,("typedef", "_TYPEDEF_")
-                                                 ,("union", "_UNION_")
-                                                 ,("unsigned", "_UNSIGNED_")
-                                                 ,("volatile", "_VOLATILE_")
-                                                 ,("void", "_VOID_")
-                                                 ,("while", "_WHILE_")])
 
 -- | From two types, one with type variables and one without (e.g. (Fn ["t0"] "t1") and (Fn [Int] Bool))
 --   create mappings that translate from the type variables to concrete types, e.g. "t0" => Int, "t1" => Bool
